[#24] Rename test_helper to capybara_helper
